+++ tests/framework/context/test_mlflow_context.py
@@ -1,11 +1,9 @@
+import pytest
 import yaml
 from kedro.framework.context import load_context
 
 from kedro_mlflow.framework.context import get_mlflow_config
-
-# def test_get_mlflow_config_outside_kedro_project(tmp_path, config_with_base_mlflow_conf):
-#     with pytest.raises(KedroMlflowConfigError, match="not a valid path to a kedro project"):
-#         get_mlflow_config(project_path=tmp_path,env="local")
+from kedro_mlflow.framework.context.config import KedroMlflowConfigError
 
 
 def _write_yaml(filepath, config):
@@ -27,7 +25,14 @@ def test_get_mlflow_config(mocker, tmp_path, config_dir):
             experiment=dict(name="fake_package", create=True),
             run=dict(id="123456789", name="my_run", nested=True),
             ui=dict(port="5151", host="localhost"),
-            hooks=dict(node=dict(flatten_dict_params=True, recursive=False, sep="-")),
+            hooks=dict(
+                node=dict(
+                    flatten_dict_params=True,
+                    recursive=False,
+                    sep="-",
+                    long_parameters_strategy="truncate",
+                )
+            ),
         ),
     )
     expected = {
@@ -37,11 +42,28 @@ def test_get_mlflow_config(mocker, tmp_path, config_dir):
         "run": {"id": "123456789", "name": "my_run", "nested": True},
         "ui": {"port": "5151", "host": "localhost"},
         "hooks": {
-            "node": {"flatten_dict_params": True, "recursive": False, "sep": "-"}
+            "node": {
+                "flatten_dict_params": True,
+                "recursive": False,
+                "sep": "-",
+                "long_parameters_strategy": "truncate",
+            }
         },
     }
     context = load_context(tmp_path)
     assert get_mlflow_config(context).to_dict() == expected
+
+
+def test_get_mlflow_config_in_uninitialized_project(mocker, tmp_path, config_dir):
+    # config_with_base_mlflow_conf is a pytest.fixture in conftest
+    mocker.patch("logging.config.dictConfig")
+    mocker.patch("kedro_mlflow.utils._is_kedro_project", return_value=True)
+
+    context = load_context(tmp_path)
+    with pytest.raises(
+        KedroMlflowConfigError, match="No 'mlflow.yml' config file found in environment"
+    ):
+        get_mlflow_config(context)
 
 
 def test_mlflow_config_with_templated_config(mocker, tmp_path, config_dir):
@@ -54,7 +76,14 @@ def test_mlflow_config_with_templated_config(mocker, tmp_path, config_dir):
             experiment=dict(name="fake_package", create=True),
             run=dict(id="123456789", name="my_run", nested=True),
             ui=dict(port="5151", host="localhost"),
-            hooks=dict(node=dict(flatten_dict_params=True, recursive=False, sep="-")),
+            hooks=dict(
+                node=dict(
+                    flatten_dict_params=True,
+                    recursive=False,
+                    sep="-",
+                    long_parameters_strategy="truncate",
+                )
+            ),
         ),
     )
 
@@ -70,7 +99,12 @@ def test_mlflow_config_with_templated_config(mocker, tmp_path, config_dir):
         "run": {"id": "123456789", "name": "my_run", "nested": True},
         "ui": {"port": "5151", "host": "localhost"},
         "hooks": {
-            "node": {"flatten_dict_params": True, "recursive": False, "sep": "-"}
+            "node": {
+                "flatten_dict_params": True,
+                "recursive": False,
+                "sep": "-",
+                "long_parameters_strategy": "truncate",
+            }
         },
     }