+++ tests/io/models/test_mlflow_model_logger_dataset.py
@@ -1,3 +1,5 @@
+from tempfile import TemporaryDirectory
+
 import mlflow
 import pandas as pd
 import pytest
@@ -27,6 +29,12 @@ def linreg_model():
 def tracking_uri(tmp_path):
     tracking_uri = (tmp_path / "mlruns").as_uri()
     return tracking_uri
+
+
+@pytest.fixture
+def tmp_folder():
+    tmp_folder = TemporaryDirectory()
+    return tmp_folder
 
 
 @pytest.fixture
@@ -275,10 +283,15 @@ def test_load_without_run_id_nor_active_run():
 
 
 def test_pyfunc_flavor_python_model_save_and_load(
-    tmp_path, tracking_uri, pipeline_ml_obj, dummy_catalog, kedro_pipeline_model
+    tmp_path,
+    tmp_folder,
+    tracking_uri,
+    pipeline_ml_obj,
+    dummy_catalog,
+    kedro_pipeline_model,
 ):
 
-    artifacts = pipeline_ml_obj.extract_pipeline_artifacts(dummy_catalog)
+    artifacts = pipeline_ml_obj.extract_pipeline_artifacts(dummy_catalog, tmp_folder)
 
     model_config = {
         "name": "kedro_pipeline_model",