+++ tests/mlflow/test_kedro_pipeline_model.py
@@ -1,4 +1,5 @@
 from pathlib import Path
+from tempfile import TemporaryDirectory
 
 import mlflow
 import pandas as pd
@@ -11,6 +12,12 @@ from sklearn.linear_model import LinearRegression
 from kedro_mlflow.io.models import MlflowModelSaverDataSet
 from kedro_mlflow.mlflow import KedroPipelineModel
 from kedro_mlflow.pipeline import pipeline_ml_factory
+
+
+@pytest.fixture
+def tmp_folder():
+    tmp_folder = TemporaryDirectory()
+    return tmp_folder
 
 
 @pytest.fixture
@@ -51,9 +58,9 @@ def pipeline_ml_obj():
     return pipeline_ml_obj
 
 
-def test_model_packaging(tmp_path, pipeline_ml_obj):
-
-    catalog = DataCatalog(
+@pytest.fixture
+def dummy_catalog(tmp_path):
+    dummy_catalog = DataCatalog(
         {
             "raw_data": MemoryDataSet(),
             "data": MemoryDataSet(),
@@ -62,12 +69,29 @@ def test_model_packaging(tmp_path, pipeline_ml_obj):
             ),
         }
     )
+    return dummy_catalog
 
-    catalog._data_sets["model"].save(2)  # emulate model fitting
 
-    artifacts = pipeline_ml_obj.extract_pipeline_artifacts(catalog)
+@pytest.mark.parametrize(
+    "copy_mode,expected",
+    [
+        (None, {"raw_data": None, "data": None, "model": None}),
+        ("assign", {"raw_data": "assign", "data": "assign", "model": "assign"}),
+        ("deepcopy", {"raw_data": "deepcopy", "data": "deepcopy", "model": "deepcopy"}),
+        ({"model": "assign"}, {"raw_data": None, "data": None, "model": "assign"}),
+    ],
+)
+def test_model_packaging_with_copy_mode(
+    tmp_path, tmp_folder, pipeline_ml_obj, dummy_catalog, copy_mode, expected
+):
 
-    kedro_model = KedroPipelineModel(pipeline_ml=pipeline_ml_obj, catalog=catalog)
+    dummy_catalog._data_sets["model"].save(2)  # emulate model fitting
+
+    artifacts = pipeline_ml_obj.extract_pipeline_artifacts(dummy_catalog, tmp_folder)
+
+    kedro_model = KedroPipelineModel(
+        pipeline_ml=pipeline_ml_obj, catalog=dummy_catalog, copy_mode=copy_mode
+    )
 
     mlflow_tracking_uri = (tmp_path / "mlruns").as_uri()
     mlflow.set_tracking_uri(mlflow_tracking_uri)
@@ -80,16 +104,31 @@ def test_model_packaging(tmp_path, pipeline_ml_obj):
         )
         run_id = mlflow.active_run().info.run_id
 
-    loaded_model = mlflow.pyfunc.load_model(
-        model_uri=(Path(r"runs:/") / run_id / "model").as_posix()
-    )
+    loaded_model = mlflow.pyfunc.load_model(model_uri=f"runs:/{run_id}/model")
+
+    # first assertion: prediction works
     assert loaded_model.predict(1) == 2
+
+    # second assertion: copy_mode works
+    actual_copy_mode = {
+        name: ds._copy_mode
+        for name, ds in loaded_model._model_impl.python_model.loaded_catalog._data_sets.items()
+    }
+
+    assert actual_copy_mode == expected
+
+
+def test_kedro_pipeline_ml_with_wrong_copy_mode_type(pipeline_ml_obj, dummy_catalog):
+    with pytest.raises(TypeError, match="'copy_mode' must be a 'str' or a 'dict'"):
+        KedroPipelineModel(
+            pipeline_ml=pipeline_ml_obj, catalog=dummy_catalog, copy_mode=1346
+        )
 
 
 # should very likely add tests to see what happens when the artifacts
 # are incorrect
 # incomplete
-# contains to input_name
+# contains no input_name
 # some memory datasets inside the catalog are persisted?
 
 
@@ -173,7 +212,7 @@ def test_model_packaging_missing_artifacts(tmp_path, pipeline_ml_obj):
         )
 
 
-def test_kedro_pipeline_ml_loading_deepcoiable_catalog(tmp_path):
+def test_kedro_pipeline_ml_loading_deepcoiable_catalog(tmp_path, tmp_folder):
 
     # create pipelien and catalog. The training will not be triggered
     def fit_fun(data):
@@ -217,7 +256,7 @@ def test_kedro_pipeline_ml_loading_deepcoiable_catalog(tmp_path):
     catalog = DataCatalog({"data": MemoryDataSet(), "model": model_dataset})
 
     kedro_model = KedroPipelineModel(pipeline_ml=ml_pipeline, catalog=catalog)
-    artifacts = ml_pipeline.extract_pipeline_artifacts(catalog)
+    artifacts = ml_pipeline.extract_pipeline_artifacts(catalog, tmp_folder)
 
     mlflow_tracking_uri = (tmp_path / "mlruns").as_uri()
     mlflow.set_tracking_uri(mlflow_tracking_uri)