+++ tests/test_utils.py
@@ -1,12 +1,9 @@
 from pathlib import Path
 
 import pytest
-import yaml
 
 from kedro_mlflow.utils import (
     KEDRO_VERSION,
-    KedroContextError,
-    _already_updated,
     _get_project_globals,
     _is_kedro_project,
     _parse_requirements,
@@ -39,35 +36,6 @@ def test_validate_project_path(project_path, project_path_expected):
 
     project_path_result = _validate_project_path(project_path)
     assert project_path_result == project_path_expected
-
-
-def test_already_updated(tmp_path):
-    def _write_yaml(filepath, config):
-        filepath.parent.mkdir(parents=True, exist_ok=True)
-        yaml_str = yaml.dump(config)
-        filepath.write_text(yaml_str)
-
-    _write_yaml(
-        tmp_path / "conf" / "base" / "mlflow.yml",
-        dict(
-            mlflow_tracking_uri=(tmp_path / "mlruns").as_posix(),
-            experiment=dict(name="Default", create=True),
-            run=dict(id=None, name=None, nested=True),
-            ui=dict(port=None, host=None),
-        ),
-    )
-    flag = _already_updated(tmp_path)
-    expected_flag = True
-
-    return flag == expected_flag
-
-
-def test_not_yet_updated(tmp_path):
-
-    flag = _already_updated(tmp_path)
-    expected_flag = False
-
-    return flag == expected_flag
 
 
 @pytest.mark.parametrize(